requirements.txt
pandas
numpy
scikit-learn
streamlit
plotly
transformers
sentence-transformers
umap-learn
hdbscan
bertopic
matplotlib
wordcloud
gensim
openpyxl